@@ -1,50 +1,81 @@
 import asyncio
-from quart import Blueprint, current_app, jsonify, request, Response
-from bento_lib.auth.quart_decorators import quart_permissions_owner
-from bento_lib.responses.quart_errors import (
-    quart_bad_request_error,
-    quart_internal_server_error
-)
+
+from fastapi import APIRouter, Form, Request, status
+from fastapi.exceptions import HTTPException
+from fastapi.responses import JSONResponse
+from pydantic import BaseModel
+from starlette.responses import Response
+from typing import Annotated
+
+from bento_lib.auth.middleware.constants import RESOURCE_EVERYTHING
 from bento_lib.types import GA4GHServiceInfo, BentoExtraServiceInfo
-from bento_drop_box_service import __version__
-from bento_drop_box_service.backend import get_backend
-from bento_drop_box_service.constants import BENTO_SERVICE_KIND, SERVICE_NAME, SERVICE_TYPE
+
+from . import __version__
+from .authz import authz_middleware
+from .backends.dependency import BackendDependency
+from .constants import BENTO_SERVICE_KIND, SERVICE_NAME, SERVICE_TYPE
+from .config import ConfigDependency
+from .logger import LoggerDependency
 
 
-drop_box_service = Blueprint("drop_box_service", __name__)
+drop_box_router = APIRouter()
+
+VIEW_PERMISSION_SET = frozenset({"view:drop_box"})
+
+authz_view_dependency = authz_middleware.dep_require_permissions_on_resource(VIEW_PERMISSION_SET)
+authz_ingest_dependency = authz_middleware.dep_require_permissions_on_resource(frozenset({"ingest:drop_box"}))
+authz_delete_dependency = authz_middleware.dep_require_permissions_on_resource(frozenset({"delete:drop_box"}))
 
 
-@drop_box_service.route("/tree", methods=["GET"])
-@quart_permissions_owner
-async def drop_box_tree() -> Response:
-    if (backend := await get_backend()) is not None:
-        return jsonify(await backend.get_directory_tree())
-
-    return quart_internal_server_error("The service source data is not configured properly")
+@drop_box_router.get("/tree", dependencies=(authz_view_dependency,))
+async def drop_box_tree(backend: BackendDependency) -> Response:
+    return JSONResponse(await backend.get_directory_tree())
 
 
-@drop_box_service.route("/objects/<path:path>", methods=["GET", "PUT"])
-@quart_permissions_owner
-async def drop_box_retrieve(path) -> Response:
-    # Werkzeug's default is to encode URL to latin1
-    # in case we have unicode characters in the filename
-    try:
-        path = path.encode('iso-8859-1').decode('utf8')
-    except UnicodeDecodeError:
-        pass
+@drop_box_router.get("/objects/{path:path}", dependencies=(authz_view_dependency,))
+async def drop_box_retrieve(path: str, backend: BackendDependency):
+    return await backend.retrieve_from_path(path)
 
-    backend = await get_backend()
 
-    if backend is None:
-        return quart_internal_server_error("The service source data is not configured properly")
+class TokenBody(BaseModel):
+    token: str
 
-    if request.method == "PUT":
-        content_length = int(request.headers.get("Content-Length", "0"))
-        if content_length == 0:
-            return quart_bad_request_error("No file provided or no/zero content length specified")
-        return await backend.upload_to_path(request, path, content_length)
+
+@drop_box_router.post("/objects/{path:path}")
+async def drop_box_retrieve_post(
+    request: Request,
+    path: str,
+    token: Annotated[str, Form()],
+    backend: BackendDependency,
+):
+    # Check the token we received in the POST body against the authorization service
+    await authz_middleware.async_check_authz_evaluate(
+        request,
+        VIEW_PERMISSION_SET,
+        RESOURCE_EVERYTHING,
+        set_authz_flag=True,
+        headers_getter=(lambda _r: {"Authorization": f"Bearer {token}"}),
+    )
 
     return await backend.retrieve_from_path(path)
+
+
+@drop_box_router.put("/objects/{path:path}", dependencies=(authz_ingest_dependency,))
+async def drop_box_upload(request: Request, path: str, backend: BackendDependency):
+    content_length = int(request.headers.get("Content-Length", "0"))
+    if content_length == 0:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST, detail="No file provided or no/zero content length specified")
+    return await backend.upload_to_path(request, path, content_length)
+
+
+@drop_box_router.delete(
+    "/objects/{path:path}",
+    status_code=status.HTTP_204_NO_CONTENT,
+    dependencies=(authz_delete_dependency,),
+)
+async def drop_box_delete(path: str, backend: BackendDependency):
+    return await backend.delete_at_path(path)
 
 
 async def _git_stdout(*args) -> str:
@@ -54,15 +85,15 @@ async def _git_stdout(*args) -> str:
     return res.decode().rstrip()
 
 
-@drop_box_service.route("/service-info", methods=["GET"])
-async def service_info() -> Response:
+@drop_box_router.get("/service-info", dependencies=[authz_middleware.dep_public_endpoint()])
+async def service_info(config: ConfigDependency, logger: LoggerDependency) -> Response:
     # Spec: https://github.com/ga4gh-discovery/ga4gh-service-info
 
     bento_info: BentoExtraServiceInfo = {
         "serviceKind": BENTO_SERVICE_KIND
     }
 
-    debug_mode = current_app.config["BENTO_DEBUG"]
+    debug_mode = config.bento_debug
     if debug_mode:
         try:
             if res_tag := await _git_stdout("describe", "--tags", "--abbrev=0"):
@@ -76,11 +107,11 @@ async def service_info() -> Response:
                 bento_info["gitCommit"] = res_commit
 
         except Exception as e:
-            current_app.logger.error(f"Error retrieving git information: {type(e).__name__}")
+            logger.error(f"Error retrieving git information: {type(e).__name__}")
 
     # Do a little type checking
     info: GA4GHServiceInfo = {
-        "id": current_app.config["SERVICE_ID"],
+        "id": config.service_id,
         "name": SERVICE_NAME,
         "type": SERVICE_TYPE,
         "description": "Drop box service for a Bento platform node.",
@@ -90,8 +121,8 @@ async def service_info() -> Response:
         },
         "contactUrl": "mailto:info@c3g.ca",
         "version": __version__,
-        "environment": "dev" if current_app.config["BENTO_DEBUG"] else "prod",
+        "environment": "dev" if debug_mode else "prod",
         "bento": bento_info,
     }
 
-    return jsonify(info)
+    return JSONResponse(info)
